+++ src/CapacitatedFacilityLocationBlock.cpp
@@ -173,9 +173,8 @@ static LinearFunction * LF( Function * f ) {
 // returns true if two vectors differ, one of them being given as a pointer
 // to an array and a subset of indices
 
-template< typename T >
-static bool is_equal( T * vec , Block::c_Subset & nms ,
-		      typename std::vector< T >::const_iterator cmp ,
+template< typename T , class It >
+static bool is_equal( T * vec , Block::c_Subset & nms , It cmp ,
 		      Block::Index n_max )
 {
  for( auto nm : nms ) {
@@ -191,9 +190,8 @@ static bool is_equal( T * vec , Block::c_Subset & nms ,
 /*--------------------------------------------------------------------------*/
 // copies one vector to a given subset of another
 
-template< typename T >
-static void copyidx( T * vec , Block::c_Subset & nms ,
-		     typename std::vector< T >::const_iterator cpy )
+template< typename T , class It >
+static void copyidx( T * vec , Block::c_Subset & nms , It cpy )
 {
  for( auto nm : nms )
   *( vec + nm ) = *(cpy++);
@@ -1992,31 +1990,38 @@ void CapacitatedFacilityLocationBlock::serialize( netCDF::NcGroup & group )
 /*--------------------------------------------------------------------------*/
 
 void CapacitatedFacilityLocationBlock::chg_facility_costs(
-				     c_CV_it NCost , Range rng ,
+				     MF_dbl_sp NCost , Range rng ,
 				     ModParam issueMod , ModParam issueAMod )
 {
  rng.second = std::min( rng.second , f_n_facilities );
  if( rng.second <= rng.first )  // nothing to change
   return;                       // cowardly (and silently) return
 
+ if( NCost.size() < rng.second - rng.first )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_facility_costs: the span is shorter "
+				"than the Range" ) );
+
+ auto NCost_it = NCost.begin();
+
  c_Index num = rng.second - rng.first;
  // TODO: if some changes are "fake", rather restrict the range
- if( std::equal( NCost , NCost + num , v_f_cost.begin() + rng.first ) )
+ if( std::equal( NCost_it , NCost_it + num , v_f_cost.begin() + rng.first ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasObj ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  std::copy( NCost , NCost + num , v_f_cost.begin() + rng.first );
+  std::copy( NCost_it , NCost_it + num , v_f_cost.begin() + rng.first );
 
   if( ( AR & FormMsk ) != KskForm )
    // since modify_coefficients owns the vector, a copy has to be made
-   get_lfo()->modify_coefficients( CVector( NCost , NCost + num ) , rng ,
+   get_lfo()->modify_coefficients( CVector( NCost_it , NCost_it + num ) , rng ,
 				   un_ModBlock( issueAMod ) );
   else {
    f_mod_skip = true;
    for( Index i = rng.first ; i < rng.second ; ++i )
-    BKB( v_Block[ i ] )->chg_profit( *(NCost++) , f_n_customers ,
+    BKB( v_Block[ i ] )->chg_profit( *(NCost_it++) , f_n_customers ,
 				     issueMod , issueAMod );
    f_mod_skip = false;
    }
@@ -2024,7 +2029,7 @@ void CapacitatedFacilityLocationBlock::chg_facility_costs(
  else
   // only change the physical representation- - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   std::copy( NCost , NCost + num , v_f_cost.begin() + rng.first );
+   std::copy( NCost_it , NCost_it + num , v_f_cost.begin() + rng.first );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2041,9 +2046,16 @@ void CapacitatedFacilityLocationBlock::chg_facility_costs(
 /*--------------------------------------------------------------------------*/
 
 void CapacitatedFacilityLocationBlock::chg_facility_costs(
-		               c_CV_it NCost , Subset && nms , bool ordered ,
+		               MF_dbl_sp NCost , Subset && nms , bool ordered ,
 			       ModParam issueMod , ModParam issueAMod )
 {
+ if( NCost.size() < nms.size() )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_facility_costs: the span is shorter "
+				"than the Subset" ) );
+
+ auto NCost_it = NCost.begin();
+
  if( nms.empty() )  // nothing to change
   return;           // cowardly (and silently) return
 
@@ -2051,23 +2063,23 @@ void CapacitatedFacilityLocationBlock::chg_facility_costs(
   throw( std::invalid_argument( "invalid facility name" ) );
 
  // TODO: eliminate from nms the "fake" changes
- if( is_equal( v_f_cost.data() , nms , NCost , f_n_facilities ) )
+ if( is_equal( v_f_cost.data() , nms , NCost_it , f_n_facilities ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasObj ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  copyidx( v_f_cost.data() , nms , NCost );
+  copyidx( v_f_cost.data() , nms , NCost_it );
 
   if( ( AR & FormMsk ) != KskForm )
    // since modify_coefficients owns both vectors, two copies are made
-   get_lfo()->modify_coefficients( CVector( NCost , NCost + nms.size() ) ,
+   get_lfo()->modify_coefficients( CVector( NCost_it , NCost_it + nms.size() ) ,
 				   Subset( nms ) , ordered ,
 				   un_ModBlock( issueAMod ) );
   else {
    f_mod_skip = true;
    for( auto i : nms )
-    BKB( v_Block[ i ] )->chg_profit( *(NCost++) , f_n_customers ,
+    BKB( v_Block[ i ] )->chg_profit( *(NCost_it++) , f_n_customers ,
 				     issueMod , un_ModBlock( issueAMod ) );
    f_mod_skip = false;
    }
@@ -2075,7 +2087,7 @@ void CapacitatedFacilityLocationBlock::chg_facility_costs(
  else
   // only change the physical representation- - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   copyidx( v_f_cost.data() , nms , NCost );
+   copyidx( v_f_cost.data() , nms , NCost_it );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2136,7 +2148,7 @@ void CapacitatedFacilityLocationBlock::chg_facility_cost( Cost NCost ,
 /*--------------------------------------------------------------------------*/
 
 void CapacitatedFacilityLocationBlock::chg_transportation_costs(
-				     c_CV_it NCost , Range rng ,
+				     MF_dbl_sp NCost , Range rng ,
 				     ModParam issueMod , ModParam issueAMod )
 {
  c_Index maxn = f_n_facilities * f_n_customers;
@@ -2144,22 +2156,29 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
  if( rng.second <= rng.first )  // nothing to change
   return;                       // cowardly (and silently) return
 
+ if( NCost.size() < rng.second - rng.first )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_transportation_costs: the span is shorter "
+				"than the Range" ) );
+
+ auto NCost_it = NCost.begin();
+
  c_Index num = rng.second - rng.first;
  // TODO: if some changes are "fake", rather restrict the range
- if( std::equal( NCost , NCost + num , v_t_cost.data() + rng.first ) )
+ if( std::equal( NCost_it , NCost_it + num , v_t_cost.data() + rng.first ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasObj ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  std::copy( NCost , NCost + num , v_t_cost.data() + rng.first );
+  std::copy( NCost_it , NCost_it + num , v_t_cost.data() + rng.first );
 
   f_mod_skip = true;
   switch( AR & FormMsk ) {
    case( StdForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - - -
     // since modify_coefficients owns the vector, a copy has to be made
-    CVector NC( NCost , NCost + num );
+    CVector NC( NCost_it , NCost_it + num );
     get_lfo()->modify_coefficients( std::move( NC ) ,
 				    Range( rng.first + f_n_facilities ,
 					   rng.second + f_n_facilities ) ,
@@ -2172,7 +2191,8 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
     Index l = f % f_n_customers;
     if( ( ( rng.second - 1 ) / f_n_customers ) == i ) {
      // the range is all inside a single facility
-     BKB( v_Block[ i ] )->chg_profits( NCost , Range( l , l + num ) ,
+     BKB( v_Block[ i ] )->chg_profits( MF_dbl_sp( & * NCost_it , num ) ,
+				       Range( l , l + num ) ,
 				       issueMod , issueAMod );
      break;
      }
@@ -2183,23 +2203,26 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
 
     // the range of the first facility does not necessarily start from 0,
     // but it surely ends at f_n_customers
-    BKB( v_Block[ i++ ] )->chg_profits( NCost , Range( l , f_n_customers ) ,
-					issueMod , iAM );
-    NCost += ( f_n_customers - l );
+    BKB( v_Block[ i++ ] )->chg_profits(
+			  MF_dbl_sp( & * NCost_it , f_n_customers - l ) ,
+			  Range( l , f_n_customers ) , issueMod , iAM );
+    NCost_it += ( f_n_customers - l );
     f += ( f_n_customers - l );
  
     // the range of all other facilities starts from 0, but it does not
     // necessarily end at f_n_customers
-    for( ; ; ++i , NCost += f_n_customers ) {
+    for( ; ; ++i , NCost_it += f_n_customers ) {
      Index nf = f + f_n_customers;
      if( nf >= rng.second ) {  // last facility
-      BKB( v_Block[ i ] )->chg_profits( NCost , Range( 0 , rng.second - f ) ,
-					issueMod , iAM );
+      BKB( v_Block[ i ] )->chg_profits(
+			  MF_dbl_sp( & * NCost_it , rng.second - f ) ,
+			  Range( 0 , rng.second - f ) , issueMod , iAM );
       break;
       }
      else {
-      BKB( v_Block[ i ] )->chg_profits( NCost , Range( 0 , f_n_customers ) ,
-					issueMod , iAM );
+      BKB( v_Block[ i ] )->chg_profits(
+			  MF_dbl_sp( & * NCost_it , f_n_customers ) ,
+			  Range( 0 , f_n_customers ) , issueMod , iAM );
       f = nf;
       }
      }
@@ -2225,7 +2248,7 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
   // only change the physical representation- - - - - - - - - - - - - - - - -
   //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   std::copy( NCost , NCost + num , v_t_cost.data() + rng.first );
+   std::copy( NCost_it , NCost_it + num , v_t_cost.data() + rng.first );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2242,9 +2265,16 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
 /*--------------------------------------------------------------------------*/
 
 void CapacitatedFacilityLocationBlock::chg_transportation_costs(
-			       c_CV_it NCost , Subset && nms , bool ordered ,
+			       MF_dbl_sp NCost , Subset && nms , bool ordered ,
 			       ModParam issueMod , ModParam issueAMod )
 {
+ if( NCost.size() < nms.size() )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_transportation_costs: the span is shorter "
+				"than the Subset" ) );
+
+ auto NCost_it = NCost.begin();
+
  if( nms.empty() )  // nothing to change
   return;           // cowardly (and silently) return
 
@@ -2256,14 +2286,14 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
 				) );
 
  // TODO: eliminate from nms the "fake" changes
- if( is_equal( v_t_cost.data() , nms , NCost , maxn ) )
+ if( is_equal( v_t_cost.data() , nms , NCost_it , maxn ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasObj ) ) {
   // change abstract and physical representation together - - - - - - - - - -
  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  copyidx( v_t_cost.data() , nms , NCost );
+  copyidx( v_t_cost.data() , nms , NCost_it );
 
   f_mod_skip = true;
   switch( AR & FormMsk ) {
@@ -2273,7 +2303,7 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
      el += f_n_facilities;
 
     // since modify_coefficients owns both vectors, copies has to be made
-    CVector NC( NCost , NCost + nms.size() );
+    CVector NC( NCost_it , NCost_it + nms.size() );
     get_lfo()->modify_coefficients( std::move( NC ) , std::move( nnms ) ,
 				    ordered , un_ModBlock( issueAMod ) );
     break;
@@ -2281,13 +2311,13 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
    case( KskForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
     // this operation is horribly complex if nms is not ordered, so order it;
     // but this means also re-ordering the values accordingly
-    auto tNC = NCost;
+    MF_dbl_sp tNC = NCost;
     CVector oNC;
     if( ! ordered ) {
      using ICPair = std::pair< Index , Cost >;
      std::vector< ICPair > tmp( nms.size() );
      for( Index i = 0 ; i < nms.size() ; ++i )
-      tmp[ i ] = std::make_pair( nms[ i ] , *(NCost++) );
+      tmp[ i ] = std::make_pair( nms[ i ] , *(NCost_it++) );
      std::sort( tmp.begin() , tmp.end() ,
 		[]( auto & a , auto & b ) { return( a.first < b.first ); } );
      oNC.resize( nms.size() );
@@ -2295,7 +2325,7 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
       nms[ i ] = tmp[ i ].first;
       oNC[ i ] = tmp[ i ].second;
       }
-     tNC = oNC.begin();
+     tNC = oNC;
      ordered = true;
      }
 
@@ -2305,8 +2335,9 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
      Subset nnms( nms );     // copy and translate names
      for( auto & el : nnms )
       el %= f_n_customers;
-     BKB( v_Block[ i ] )->chg_profits( tNC , std::move( nnms ) , true ,
-				       issueMod , issueAMod );
+     const auto nn = nnms.size();
+     BKB( v_Block[ i ] )->chg_profits( tNC.first( nn ) , std::move( nnms ) ,
+				       true , issueMod , issueAMod );
      break;
      }
 
@@ -2325,12 +2356,13 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
      for( auto & el : nnms )
       el %= f_n_customers;
 
-     BKB( v_Block[ i ] )->chg_profits( tNC , std::move( nnms ) , true ,
-				       issueMod , iAM );
+     const auto nn = nnms.size();
+     BKB( v_Block[ i ] )->chg_profits( tNC.first( nn ) , std::move( nnms ) ,
+				       true , issueMod , iAM );
      if( eit == nms.end() )
       break;
 
-     tNC += std::distance( bit , eit );
+     tNC = tNC.subspan( std::distance( bit , eit ) );
      bit = eit;
      }
 
@@ -2355,7 +2387,7 @@ void CapacitatedFacilityLocationBlock::chg_transportation_costs(
   // only change the physical representation- - - - - - - - - - - - - - - - -
   //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   copyidx( v_t_cost.data() , nms , NCost );
+   copyidx( v_t_cost.data() , nms , NCost_it );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2441,22 +2473,29 @@ void CapacitatedFacilityLocationBlock::chg_transportation_cost( Cost NCost ,
 /*--------------------------------------------------------------------------*/
 
 void CapacitatedFacilityLocationBlock::chg_facility_capacities(
-				     c_DV_it NCap , Range rng ,
+				     MF_dbl_sp NCap , Range rng ,
 				     ModParam issueMod , ModParam issueAMod )
 {
  rng.second = std::min( rng.second , f_n_facilities );
  if( rng.second <= rng.first )  // nothing to change
   return;                       // cowardly (and silently) return
 
+ if( NCap.size() < rng.second - rng.first )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_facility_capacities: the span is shorter "
+				"than the Range" ) );
+
+ auto NCap_it = NCap.begin();
+
  Index num = rng.second - rng.first;
  // TODO: if some changes are "fake", rather restrict the range
- if( std::equal( NCap , NCap + num , v_capacity.begin() + rng.first ) )
+ if( std::equal( NCap_it , NCap_it + num , v_capacity.begin() + rng.first ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasCapCns ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  std::copy( NCap , NCap + num , v_capacity.begin() + rng.first );
+  std::copy( NCap_it , NCap_it + num , v_capacity.begin() + rng.first );
 
   // if appropriate, open a new channel to bunch up all abstract Modification
   not_ModBlock( issueAMod );
@@ -2469,20 +2508,21 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
    case( StdForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( Index i = rng.first ; i < rng.second ; ++i )
      LF( v_cap[ i ].get_function()
-	 )->modify_coefficient( f_n_customers , - *(NCap++) , iAM );
+	 )->modify_coefficient( f_n_customers , - *(NCap_it++) , iAM );
     break;
     }
    case( KskForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( Index i = rng.first ; i < rng.second ; ++i )
      BKB( v_Block[ i ]
-	  )->chg_weight( - *(NCap++) , f_n_customers , issueMod , iAM );
+	  )->chg_weight( - *(NCap_it++) , f_n_customers , issueMod , iAM );
     break;
     }
    case( FlwForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
-    MCFB( v_Block[ 1 ] )->chg_ucaps( NCap , rng , issueMod , iAM );
+    MCFB( v_Block[ 1 ] )->chg_ucaps( NCap.first( num ) , rng , issueMod ,
+				     iAM );
     for( Index i = rng.first ; i < rng.second ; ++i )
      LF( v_cap[ i ].get_function()
-	 )->modify_coefficient( 1 , - *(NCap++) , iAM );
+	 )->modify_coefficient( 1 , - *(NCap_it++) , iAM );
     break;
     }
    case( BenForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -2491,7 +2531,8 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
     // namespace (no parent): the channel of iAM was opened on *this*, so we
     // strip it via par2mod() before forwarding to mcfb. MCFBlock::chg_ucaps
     // will open its own internal channel if num > 1.
-    mcfb->chg_ucaps( NCap , rng , issueMod , Observer::par2mod( iAM ) );
+    mcfb->chg_ucaps( NCap.first( num ) , rng , issueMod ,
+		     Observer::par2mod( iAM ) );
     // update A_{i,i} = v_capacity[ i ] in the BendersBFunction mapping;
     // the v_capacity vector was just updated above so re-read from it
     for( Index i = rng.first ; i < rng.second ; ++i ) {
@@ -2511,7 +2552,7 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
  else
   // only change the physical representation- - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   std::copy( NCap , NCap + num , v_capacity.begin() + rng.first );
+   std::copy( NCap_it , NCap_it + num , v_capacity.begin() + rng.first );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2531,9 +2572,16 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
 /*--------------------------------------------------------------------------*/
 
 void CapacitatedFacilityLocationBlock::chg_facility_capacities(
-			        c_DV_it NCap , Subset && nms , bool ordered ,
+			        MF_dbl_sp NCap , Subset && nms , bool ordered ,
 				ModParam issueMod , ModParam issueAMod )
 {
+ if( NCap.size() < nms.size() )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_facility_capacities: the span is shorter "
+				"than the Subset" ) );
+
+ auto NCap_it = NCap.begin();
+
  if( nms.empty() )  // nothing to change
   return;           // cowardly (and silently) return
 
@@ -2541,13 +2589,13 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
   throw( std::invalid_argument( "invalid facility name" ) );
 
  // TODO: eliminate from nms the "fake" changes
- if( is_equal( v_capacity.data() , nms , NCap , f_n_facilities ) )
+ if( is_equal( v_capacity.data() , nms , NCap_it , f_n_facilities ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasCapCns ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  copyidx( v_capacity.data() , nms , NCap );
+  copyidx( v_capacity.data() , nms , NCap_it );
 
   // if appropriate, open a new channel to bunch up all abstract Modification
   Index num = nms.size() + ( ( AR & FormMsk ) == FlwForm ? 1 : 0 );
@@ -2559,21 +2607,22 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
    case( StdForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( auto i : nms )
      LF( v_cap[ i ].get_function()
-	 )->modify_coefficient( f_n_customers , - *(NCap++) , iAM );
+	 )->modify_coefficient( f_n_customers , - *(NCap_it++) , iAM );
     break;
     }
    case( KskForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( auto i : nms )
      BKB( v_Block[ i ]
-	  )->chg_weight( - *(NCap++) , f_n_customers , issueMod , iAM );
+	  )->chg_weight( - *(NCap_it++) , f_n_customers , issueMod , iAM );
     break;
     }
    case( FlwForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
-    MCFB( v_Block[ 1 ] )->chg_ucaps( NCap , Subset( nms ) , ordered ,
+    MCFB( v_Block[ 1 ] )->chg_ucaps( NCap.first( nms.size() ) ,
+				     Subset( nms ) , ordered ,
 				     issueMod , iAM );
     for( auto i : nms )
      LF( v_cap[ i ].get_function()
-	 )->modify_coefficient( 1 , - *(NCap++) , iAM );
+	 )->modify_coefficient( 1 , - *(NCap_it++) , iAM );
     break;
     }
    case( BenForm ): {  // - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -2581,8 +2630,8 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
     // mcfb is the hidden inner Block of f_BF and lives in its own channel
     // namespace (no parent): the channel of iAM was opened on *this*, so we
     // strip it via par2mod() before forwarding to mcfb.
-    mcfb->chg_ucaps( NCap , Subset( nms ) , ordered , issueMod ,
-                     Observer::par2mod( iAM ) );
+    mcfb->chg_ucaps( NCap.first( nms.size() ) , Subset( nms ) , ordered ,
+                     issueMod , Observer::par2mod( iAM ) );
     for( auto i : nms ) {
      BendersBFunction::RealVector Ai( f_n_facilities , 0 );
      Ai[ i ] = v_capacity[ i ];
@@ -2599,7 +2648,7 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacities(
  else
   // only change the physical representation- - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   copyidx( v_capacity.data() , nms , NCap );
+   copyidx( v_capacity.data() , nms , NCap_it );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2685,22 +2734,29 @@ void CapacitatedFacilityLocationBlock::chg_facility_capacity( Demand NCap ,
 
 /*--------------------------------------------------------------------------*/
 
-void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
+void CapacitatedFacilityLocationBlock::chg_customer_demands( MF_dbl_sp NDem ,
 			 Range rng , ModParam issueMod , ModParam issueAMod )
 {
  rng.second = std::min( rng.second , f_n_customers );
  if( rng.second <= rng.first )  // nothing to change
   return;                       // cowardly (and silently) return
 
+ if( NDem.size() < rng.second - rng.first )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_customer_demands: the span is shorter "
+				"than the Range" ) );
+
+ auto NDem_it = NDem.begin();
+
  c_Index num = rng.second - rng.first;
  // TODO: if some changes are "fake", rather restrict the range
- if( std::equal( NDem , NDem + num , v_demand.begin() + rng.first ) )
+ if( std::equal( NDem_it , NDem_it + num , v_demand.begin() + rng.first ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasSatCns ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  std::copy( NDem , NDem + num , v_demand.begin() + rng.first );
+  std::copy( NDem_it , NDem_it + num , v_demand.begin() + rng.first );
 
   // if appropriate, open a new channel to bunch up all abstract Modification
   Index nc = ( ( AR & FormMsk ) == FlwForm ) ? 0 : f_n_facilities;
@@ -2712,12 +2768,13 @@ void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
    case( StdForm ):    // - - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( auto & capi : v_cap )
      LF( capi.get_function()
-	 )->modify_coefficients( DVector( NDem , NDem + num ) , rng , iAM );
+	 )->modify_coefficients( DVector( NDem_it , NDem_it + num ) , rng , iAM );
     break;
 
    case( KskForm ):    // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( auto bi : v_Block )
-     BKB( bi )->chg_weights( NDem , rng , issueMod , iAM );
+     BKB( bi )->chg_weights( NDem.first( rng.second - rng.first ) , rng ,
+			     issueMod , iAM );
 
     break;
 
@@ -2739,7 +2796,7 @@ void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
  else
   // only change the physical representation- - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   std::copy( NDem , NDem + num , v_demand.begin() + rng.first );
+   std::copy( NDem_it , NDem_it + num , v_demand.begin() + rng.first );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;
@@ -2755,10 +2812,17 @@ void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
 
 /*--------------------------------------------------------------------------*/
 
-void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
+void CapacitatedFacilityLocationBlock::chg_customer_demands( MF_dbl_sp NDem ,
 			            Subset && nms , bool ordered ,
 				    ModParam issueMod , ModParam issueAMod )
 {
+ if( NDem.size() < nms.size() )
+  throw( std::invalid_argument( "CapacitatedFacilityLocationBlock::"
+				"chg_customer_demands: the span is shorter "
+				"than the Subset" ) );
+
+ auto NDem_it = NDem.begin();
+
  if( nms.empty() )  // nothing to change
   return;           // cowardly (and silently) return
 
@@ -2766,13 +2830,13 @@ void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
   throw( std::invalid_argument( "invalid customer name" ) );
 
  // TODO: eliminate from nms the "fake" changes
- if( is_equal( v_demand.data() , nms , NDem , f_n_customers ) )
+ if( is_equal( v_demand.data() , nms , NDem_it , f_n_customers ) )
   return;  // actually nothing changes, avoid issuing the Modification
 
  if( not_dry_run( issueAMod ) && ( AR & HasSatCns ) ) {
   // change abstract and physical representation together - - - - - - - - - -
   // in the meantime, if so instructed also issue abstract Modification
-  copyidx( v_demand.data() , nms , NDem );
+  copyidx( v_demand.data() , nms , NDem_it );
 
   // if appropriate, open a new channel to bunch up all abstract Modification
   Index nc = ( ( AR & FormMsk ) == FlwForm ) ? 0 : f_n_facilities;
@@ -2784,14 +2848,14 @@ void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
    case( StdForm ):    // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( auto & capi : v_cap )
      LF( capi.get_function()
-	 )->modify_coefficients( DVector( NDem , NDem + nms.size() ) ,
+	 )->modify_coefficients( DVector( NDem_it , NDem_it + nms.size() ) ,
 				 Subset( nms ) , ordered , iAM );
     break;
  
    case( KskForm ):    // - - - - - - - - - - - - - - - - - - - - - - - - - -
     for( auto bi : v_Block )
-     BKB( bi )->chg_weights( NDem , Subset( nms ) , ordered ,
-			     issueMod , iAM );
+     BKB( bi )->chg_weights( NDem.first( nms.size() ) , Subset( nms ) ,
+			     ordered , issueMod , iAM );
 
     break;
 
@@ -2812,7 +2876,7 @@ void CapacitatedFacilityLocationBlock::chg_customer_demands( c_DV_it NDem ,
  else
   // only change the physical representation- - - - - - - - - - - - - - - - -
   if( not_dry_run( issueMod ) )
-   copyidx( v_demand.data() , nms , NDem );
+   copyidx( v_demand.data() , nms , NDem_it );
 
  f_cond_lower = NAN;  // reset conditional bounds
  f_cond_upper = NAN;